+++ CHANGELOG.md
@@ -3,11 +3,15 @@
 
 1. `Signal` now offers a non-serializing variant and a reentrant variant for advanced users. (#797)
 
+   The input observer of these variants assume that mutual exclusion has already been enforced among its callers.
+
    You can create these variants through four `Signal` static methods: `nonSerializing(_:)`, `nonSerializingPipe(_:)`, `reentrant(_:)` and `reentrantPipe(_:)`. These would be adopted by ReactiveCocoa UIKit bindings to improve interoperability with Loop, to tackle some legitimate recursive delivery scenarios (e.g. around first responder management), and also to reduce fine-grained locking in ReactiveCocoa.
 
    Note that the default behavior of `Signal` has not been changed — event serialization remains the default behavior.
 
-1. `Signal` now has reduced fine-grained locking by adopting `nonSerializing(_:)` internally. (#797)
+1. `SignalProducer` now similarly offers `nonSerializing(_:)`. (#797)
+
+1. `Signal` and Properties now use fewer locks, which should translate into minor performance improvements. (#797)
 
 1. Fix a debug assertion in `Lock.try()` that could be raised in earlier OS versions (< iOS 10.0, < macOS 10.12). (#747, #788)